\name{affineModelR-package}
\alias{affineModelR-package}
\alias{affineModelR}
\docType{package}
\title{
\packageTitle{affineModelR}
}
\description{
\packageDescription{affineModelR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{affineModelR}
\packageIndices{affineModelR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{affineModelR}

Maintainer: \packageMaintainer{affineModelR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
