% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affineSimulate.R
\name{affineSimulate}
\alias{affineSimulate}
\title{This function simulates paths from a stochastic volatility model with three volatility factors. One of the factors and the underlying asset prices can co-jump. The jump specification follows Duffie, Pan and Singleton (2000), p. 1361.}
\usage{
affineSimulate(paramsList, N.factors = 3, t.days = 1, t.freq = 1/78,
  freq.subdiv = 24, rng.seed = 42, init.vals = NULL, rf.rate = 0,
  jumpGeneratorPtr = getPointerToGenerator(fstr = "expNormJumpTransform"),
  jumpTransformPtr = getPointerToJumpTransform(fstr =
  "expNormJumpTransform")$TF, mod.type = "standard",
  specMaker = ODEstructsForSim, nrepl = 1, ...)
}
\arguments{
\item{paramsList}{A list containing fields \code{P} and \code{Q}, containing, respectively, the P and Q parameters.}

\item{N.factors}{number of stochastic volatility factors (stock is not a factor)}

\item{t.days}{Number of days to simulate}

\item{t.freq}{Number of intraday prices to simulate}

\item{freq.subdiv}{Number of periods each segment should be broken into when simulating}

\item{rng.seed}{The random seed used for generations}

\item{init.vals}{Initial values for the of the stock/volatility to start from (in case we want to continue a given path, if NULL stock will be started from 1 and the volatilities from stationary distribution.)}

\item{rf.rate}{Assumed constant risk-free rate (annualized, continuously compounded)}

\item{specMaker}{function that takes arguments \code{params.P}, \code{params.Q}, \code{jumpTransform}, \code{N.factors}, \code{rf.rate}, \code{mod.type} and potentially more via the ellipsis operator (\code{...}) and returns the DPS matrices. See documentation of \link{ODEstructs}.}

\item{jumpTransform}{The function that calculates the jump Laplace transform}
}
\value{
A list containing the simulated stock & volatility paths + stock price jump times + a time index
}
\description{
This function simulates paths from a stochastic volatility model with three volatility factors. One of the factors and the underlying asset prices can co-jump. The jump specification follows Duffie, Pan and Singleton (2000), p. 1361.
}

