% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpTransforms.R
\name{expNormJumpTransform}
\alias{expNormJumpTransform}
\title{non-vectorized jumpTransform of the DPS exponential-normal mixture}
\usage{
expNormJumpTransform(uu, jmp)
}
\arguments{
\item{uu}{(N.factors+1) length vector where the jump transform should be evaluated}

\item{jmp}{a structure describing the jump parameters (muY/sigmaY = mean/sd of log stock jump)}
}
\value{
scalar value corresponding to the jump transform evaluated at the given frequency
}
\description{
non-vectorized jumpTransform of the DPS exponential-normal mixture
}
\references{
Duffie-Pan-Singleton (2000)
}
