% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finiteDifferenceDerivatives.R
\name{fdCoefficients}
\alias{fdCoefficients}
\title{Finite difference coefficients.}
\usage{
fdCoefficients(p, k)
}
\arguments{
\item{p}{Derivative order.}

\item{k}{Desired accuracy.}

\item{h}{Grid step size.}

\item{f0}{Value of function at differentiation point.}

\item{fVal}{Values of function at points that will be used for approximation.}
}
\value{
\code{fdCoefficients} Vector of coefficients for linear combination of function values that will give the desired finite-difference approximation. \cr \cr \code{fdDerivative} numerical derivative based on vector of function values.
}
\description{
Calculation of coefficients for finite-difference approximation of derivative of order p with error of order O(h^(p+k)), given a vector of function values.
}
\details{
If p+k is odd, the finite difference scheme will be asymmetric, with one more point on the negative side.
}
