% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makePositiveDefinite}
\alias{makePositiveDefinite}
\title{Make Positive Definite}
\usage{
makePositiveDefinite(semiDefMat, relEig = 1e-06)
}
\arguments{
\item{semiDefMat}{semi-pos-def matrix, symmetric}

\item{relEig}{relative range of eigenvalues taken to define the pd matrix}
}
\value{
matrix size of semiDefMat
}
\description{
Take a semi-positive definite matrix and use the eigenvalue decomposition to obtain the nearest positive definite matrix
}

