% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEstructs.R
\name{ODEstructs}
\alias{ODEstructs}
\title{ODE specification structures}
\usage{
ODEstructs(params, jumpTransform = NULL, mkt, N.factors)
}
\arguments{
\item{params}{A structure containing the volatility factor + jump descriptions.}

\item{jumpTransform}{String, the function that calculates the jump transform.}

\item{N.factors}{The number of stochastic volatility factors (stock is not a factor).}
}
\value{
Return a list with elements K0, l0, K1, l1, H1, see DPS (2000) for their meaning
}
\description{
This function translates the parameters lists (see \code{jumpDiffusionODEs}) to model specification matrices and vectors for ODE solution. Expand this function if you want to add other specifications.
}
