% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specMakers.R
\name{ODEstructsForSim}
\alias{ODEstructsForSim}
\title{Calculate vectors and matrices for propagation equations in a N-factor model with co-jumps in the underlying and the first volatility factor}
\usage{
ODEstructsForSim(params.P = NULL, params.Q,
  jumpTransformPointer = getPointerToJumpTransform(fstr =
  "expNormJumpTransform"), N.factors, rf.rate = 0, mod.type = "standard")
}
\arguments{
\item{params.P}{A structure containing the volatility factor + jump descriptions under measure P}

\item{params.Q}{A structure containing the volatility factor + jump descriptions under measure Q}

\item{N.factors}{The number of stochastic volatility factors (stock is not a factor)}

\item{rf.rate}{risk-free rate}

\item{jumpTransform}{Pointer to the basic jumpTransform c++ function. Via this mechanism it is easy for the user to provide their external jump transform functions for both simulation and ODE calculations.}
}
\value{
Return a list with elements \code{terms.dt}, \code{terms.vdt}, \code{terms.vdW}, \code{terms.vdWort}, \code{vterms.dt}, \code{vterms.vdt}, \code{vterms.vdW}, \code{jmpPar}, \code{intensity.terms}. These are all elements for simulating the N-factor model.
}
\description{
Calculate vectors and matrices for propagation equations in a N-factor model with co-jumps in the underlying and the first volatility factor
}
